requirements.txt

python-multipart
pytesseract
pdf2image
python-dotenv
pillow
python-magic
scikit-learn
nltk
# Core dependencies
fastapi
uvicorn
python-dotenv
motor

aio_pika
# Messaging
# Utilities
tenacity
aiofiles

# AI and Processing
google-generativeai
PyPDF2
docx2txt
pytesseract
pdf2image
config
# Development
pytest
pytest-asyncio
httpx
